Add stor.utils.is_writeable. (#29)
The goal of stor.utils.is_writeable is to be able to determine whether we have write access to the specified path, either local, S3 or Swift. This is useful, because using this we'd be able to check for write permissions ahead-of-time, before generating objects for upload, which might take significant amount of time.

Check the docstring and tests for more details.