Do not generate buffers just to close them.
Previously we were creating buffers in the ``close()`` method in order
to call ``close()`` on them and/or causing uploads in the close method
in an attempt to upload an empty file. For strict compatibility with
POSIX, we probably should create an empty file if we open but do
not close an object, but this means we can potentially generate an error
in a close method in an edge case that isn't that fun.

This is not an API-breaking change, because previously we checked for
buffer.tell() in flush, so no empty files are created.