TQDM progress logger
Replace the existing progress logger with tqdm, a library that makes
elegant progress bars that calculate rate of transfer, time remaining,
and also look nice on the CLI (i.e., stays on one line and doesn't gum
up the entire works). A future enhancement would be to make the progress
logger configurable, but I figure for now just showing a better one is a
bonus.

How this was tested:

1. Tested stor cp -r and rm -r with various combinatiosn
2. Ensure passes all unit tests.

Sem-Ver: feature