Fail earlier when `open` called on dx project path
Raise an exception directly in `stor.open` if we have a project path.

Previously, we were raising an exception at `write_object`;
however, if you actually attempted to write data to the path, this would
result in an exception in the `__del__` method as `stor` tried to flush
data.

This also helps reduce noise from the test suite.

Doc note