Generate a useful exception on ImportError
Rather than leave the user hanging, show something helpful to the user
so they can resolve the situation by installing an additional package.

Left this as a separate function so it's easy to test the exception
generation fucntion doesn't break without needing to do painful testing
contortions. (which should be caught later on by test matrix)